fix the backpressures(if file too big will make a memory issue) in the uploader app

+++ networking/uploader-app/server.js
@@ -3,22 +3,36 @@ const fs = require("fs/promises");
 
 const server = net.createServer(() => {})
 
-let fileHandle, fileStream
+let fileHandle, fileWriteStream
 
 server.on('connection', (socket) => {
     console.log("New connection!")
 
     socket.on("data", async (data) => {
-        fileHandle = await fs.open(`storage/test.txt`, "w")
-        fileStream = fileHandle.createWriteStream()
+        if(!fileHandle) {     // If file handle doesn't exist, create one and start writing to a file
+            socket.pause()
+            fileHandle = await fs.open(`storage/test.txt`, "w")
+            fileWriteStream = fileHandle.createWriteStream() //stream to write to
+    
+            // Writing to our dest file
+            fileWriteStream.write(data)
 
-        // Writing to our dest file
-        fileStream.write(data)
+            socket.resume(); // resume receiving data from the client
+            fileWriteStream.on('drain', () => {
+                socket.resume();
+            });
+        }else {     // If file handle exists, continue writing to the existing file
+            if (!fileWriteStream.write(data)) {
+                socket.pause();
+            }
+        }
     })
-
+    // This end event happens when the client.js file ends the socket
     socket.on('end', () => {
+        fileHandle.close();
+        fileHandle = undefined;
+        fileWriteStream = undefined;
         console.log('connection endedd!')
-        fileHandle.close()
     })
 })
